Updated match proteins builder with better decoy detection.
Ensured no decoy protein sequences are included, even if some of the
headers associated with that sequence aren't labelled as decoys. This
can theoretically happen if a FASTA header is pre-fixed with "DECOY_",
but has separate headers separated by control-a characters. None of the
other headers would appear to be decoy headers.